+++ packages/pug/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/pug
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.3.13
+
 ## 0.1.13
 
 ### Patch Changes